requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
pydantic==1.10.11
python-dotenv==1.0.0
numpy==1.26.5
pandas==2.2.2
scikit-learn==1.3.2
torch==2.2.2
transformers==4.40.0
sentence-transformers==2.2.2
ftfy==6.1.1
opencv-python==4.8.1.78
hdbscan==0.8.34
qdrant-client==1.0.0
psycopg2-binary==2.9.10
pytest==7.4.2
Pillow==10.0.0
aiohttp==3.9.4